router-policy-to-config/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "router-policy-to-config"
version = "0.1.0"
description = "AI-assisted copilot for router configuration - transform YAML policies into RouterOS and OpenWrt configs"
authors = [{name = "run-as-daemon", email = "contact@run-as-daemon.ru"}]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0",
  "jsonschema>=4.17",
  "typer>=0.9.0",
  "rich>=13.0",
  "requests>=2.31",
  "jinja2>=3.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.1.8",
  "black>=23.11",
  "isort>=5.12",
  "bandit>=1.7",
  "pip-audit>=2.5",
  "mypy>=1.7",
  "pytest-cov>=4.1"
]

[project.scripts]
router-policy = "router_policy_to_config.cli:app"

[tool.ruff]
line-length = 120
extend-select = ["I"]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-v --cov=router_policy_to_config"
python_files = ["test_*.py"]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
